requirements.txt
argcomplete==3.2.1
asgiref==3.7.2
boto3==1.34.7
botocore==1.34.7
certifi==2023.11.17
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.37.0
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==5.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-s3-storage==0.15.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
durationpy==0.6
hjson==3.1.0
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
mysqlclient==2.2.1
oauthlib==3.2.2
placebo==0.9.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stripe==7.9.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.3
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
Werkzeug==3.0.1
